% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_data_flag}
\alias{example_data_flag}
\title{Derive flag data}
\usage{
example_data_flag(active_filtered_flags)
}
\arguments{
\item{active_filtered_flags}{(\code{character}) vector that
may only contain \code{error}, \code{warn}, \code{info}}
}
\description{
Derive flag data
}
