% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_data_plot}
\alias{example_data_plot}
\title{Derive economics data}
\usage{
example_data_plot()
}
\value{
economics data from ggplot2
}
\description{
Derive economics data
}
\author{
Sebastian Engel-Wolf\url{mailto:sebastian@mail-wolf.de}
}
