% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filteringapp-package.R
\docType{package}
\name{filteringapp-package}
\alias{filteringapp}
\alias{filteringapp-package}
\title{filteringapp: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\details{
This package is used to generate a data filtering app
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
