% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterNumeric.R
\name{FilterNumeric}
\alias{FilterNumeric}
\title{Numeric Filter}
\description{
Numeric Filter

Numeric Filter
}
\section{Super class}{
\code{\link[filteringapp:GenericFilter]{filteringapp::GenericFilter}} -> \code{FilterNumeric}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FilterNumeric$new()}}
\item \href{#method-get_ui}{\code{FilterNumeric$get_ui()}}
\item \href{#method-get_server}{\code{FilterNumeric$get_server()}}
\item \href{#method-clone}{\code{FilterNumeric$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructor for a numeric filter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterNumeric$new(active = FALSE, name = NULL, min, max)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{active}}{(\code{logical}) Whether the
filter should be active per-se}

\item{\code{name}}{(\code{character}) Name of
the Filter. normally the column name
of the column the filter is applied to}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ui"></a>}}
\if{latex}{\out{\hypertarget{method-get_ui}{}}}
\subsection{Method \code{get_ui()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterNumeric$get_ui()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_server"></a>}}
\if{latex}{\out{\hypertarget{method-get_server}{}}}
\subsection{Method \code{get_server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterNumeric$get_server()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterNumeric$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
