% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_value_limit.R
\name{module_ui_value_limit}
\alias{module_ui_value_limit}
\title{Module for limited values}
\usage{
module_ui_value_limit(id, value)
}
\arguments{
\item{id}{(\code{character}) UI ID}

\item{value}{(\code{numeric}) default value of the input field

renders an input field with an additional error
field to show if the numeric value is inside
or outside of a range of desired values

for a useful example see \link{module_value_limit}}
}
\value{
\link[shiny]{tagList} with a numeric input field
}
\description{
Module for limited values
}
\seealso{
module_value_limit
}
\author{
Sebastian Engel-Wolf \link{mailto:sebastian@mail-wolf.de}
}
