% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_value_limit.R
\name{module_value_limit}
\alias{module_value_limit}
\title{Module for limited values - server}
\usage{
module_value_limit(input, output, session, border_value)
}
\arguments{
\item{input}{(from shiny)}

\item{output}{(from shiny)}

\item{session}{(from shiny)}

\item{border_value}{(\code{numeric}) vector of length 2 that
defines the lower and upper limit of values allowed
inside the input}
}
\value{
shiny module server function ony functional with
\link{module_ui_value_limit}
}
\description{
Module for limited values - server
}
\examples{

ui <- shiny::fluidPage(
  tags$span("Only works with values between 1 and 6"),
  module_ui_value_limit(id = "noID", value = 5),
  textOutput(outputId = "out")
)
server <- function(input, output, session) {
  val <- reactive({
    shiny::callModule(
     module = module_value_limit,
     id = "noID",
     border_value = c(1, 6)
    )
  })
  output$out <- shiny::renderText({val()})
}

if (interactive()) {
  shiny::shinyApp(ui, server)
}

}
\seealso{
module_ui_value_limit
}
\author{
Sebastian Engel-Wolf \link{mailto:sebastian@mail-wolf.de}
}
